requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altair==4.0.0
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
attrs==24.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dataclasses-json==0.5.14
distro==1.9.0
entrypoints==0.4
exceptiongroup==1.2.2
faiss-cpu==1.7.4
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
groq==0.11.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.14.1
idna==3.10
importlib_metadata==8.5.0
InstructorEmbedding==1.0.1
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.4
langchain-community==0.3.3
langchain-core==0.3.13
langchain-groq==0.2.0
langchain-text-splitters==0.3.0
langsmith==0.1.137
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
nltk==3.9.1
numexpr==2.10.1
numpy==1.26.4
openai==0.27.6
openapi-schema-pydantic==1.2.4
orjson==3.10.10
packaging==24.1
pandas==2.2.3
pillow==11.0.0
pinecone==5.3.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
propcache==0.2.0
protobuf==3.20.3
pyarrow==18.0.0
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
Pympler==1.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.2
pywin32==308
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.3
rpds-py==0.20.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
semver==3.0.2
sentence-transformers==2.2.2
sentencepiece==0.2.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
streamlit==1.18.1
sympy==1.13.1
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.4.0
tokenizers==0.13.3
toml==0.10.2
toolz==1.0.0
torch==2.5.0
torchvision==0.20.0
tornado==6.4.1
tqdm==4.66.6
transformers==4.31.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
validators==0.34.0
watchdog==5.0.3
yarl==1.16.0
zipp==3.20.2